[FIX] parse: resolve invalid targets
In some cases, the as3 compiler emits Jump instruction in dead code (so
it's never reachable and it's ok to have an invalid instruction there).
In the future, it might be a good idea to add a stricter check on the
instruction type to only allow Jumps to have this behaviour.

Dead code analysis would also fix this issue.