[IMP] abcfile: split abc data to its own struct
This simplify borrow checking access when calling methods.
In many cases, it's needed to borrow both something from the constant
pool and something from the Abc data. For example, renaming a class
requires to borrow a Class and edit the constant pool of strings. Using
a method on the AbcFile will borrow both, and prevent doing a simple
rename without extra borrow logic.